// RESPONSIVE WEB DEVELOPMENT/src/interfaces/paciente.interfaces.ts
export interface iUpdatePatient {
    nome: string;
    email: string;
    senha: string;
    cpf: string;
    idade: string;
    endereco: string;
}


export interface iExam {
    leukocytes: string;
    platelets: string;
    meanPlateletVolume: string;
    eosinophils: string;
    proteinaCReativa: string;
}